requirements.txt
Django

mozilla-django-oidc
django-crispy-forms
django-ajax-datatable
pytz

django-autosecretkey

# For MySQL

mysqlclient